# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface Contact {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  organization?: string;
  /** e.g. band_director, event_coordinator, venue, other */
  type: string;
  notes?: string;
  lastOutreachAt?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContactBody {
  name: string;
  email?: string;
  phone?: string;
  organization?: string;
  type: string;
  notes?: string;
}

export interface Outreach {
  id: number;
  contactId: number;
  eventId?: number;
  /** e.g. email, phone, in-person, text */
  method: string;
  notes?: string;
  outreachAt: string;
  createdAt: string;
}

export interface CreateOutreachBody {
  method: string;
  notes?: string;
  eventId?: number;
  outreachAt?: string;
}

export interface Event {
  id: number;
  title: string;
  /** student_showcase, community_event, recital, open_mic, other */
  type: string;
  /** planning, confirmed, completed, cancelled */
  status: string;
  description?: string;
  location?: string;
  startDate?: string;
  endDate?: string;
  googleCalendarEventId?: string;
  /** Tag for website calendar script integration */
  calendarTag?: string;
  /** Are we being paid for sound services? */
  isPaid?: boolean;
  /** Cost in dollars (what we pay) */
  cost?: number;
  /** Revenue in dollars (what we receive) */
  revenue?: number;
  notes?: string;
  signupToken?: string;
  signupDeadline?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEventBody {
  title: string;
  type: string;
  status: string;
  description?: string;
  location?: string;
  startDate?: string;
  endDate?: string;
  googleCalendarEventId?: string;
  calendarTag?: string;
  isPaid?: boolean;
  cost?: number;
  revenue?: number;
  notes?: string;
  signupDeadline?: string;
}

export interface AddEventContactBody {
  contactId: number;
  role?: string;
}

export interface Employee {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  /** staff, intern */
  role: string;
  isActive: boolean;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEmployeeBody {
  name: string;
  email?: string;
  phone?: string;
  role: string;
  isActive?: boolean;
  notes?: string;
}

export interface EventEmployee {
  id: number;
  eventId: number;
  employeeId: number;
  employeeName?: string;
  employeeRole?: string;
  /** Role at this specific event */
  role?: string;
  pay?: number;
  notes?: string;
}

export interface AddEventEmployeeBody {
  employeeId: number;
  role?: string;
  pay?: number;
  notes?: string;
}

export interface EventSignup {
  id: number;
  eventId: number;
  name: string;
  email?: string;
  phone?: string;
  role?: string;
  notes?: string;
  /** pending, confirmed, declined */
  status: string;
  createdAt: string;
}

export interface SubmitSignupBody {
  name: string;
  email?: string;
  phone?: string;
  role?: string;
  notes?: string;
}

export interface NotifyResponse {
  success: boolean;
  message: string;
  signupUrl?: string;
}

export interface SignupPageInfo {
  eventTitle: string;
  eventType: string;
  startDate?: string;
  endDate?: string;
  location?: string;
  description?: string;
  signupDeadline?: string;
}

export interface DashboardStats {
  totalContacts: number;
  upcomingEvents: number;
  totalEmployees: number;
  pendingSignups: number;
  recentOutreach: Outreach[];
  upcomingEventsList: Event[];
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type ListContactsParams = {
  search?: string;
  type?: string;
};

export type ListEventsParams = {
  status?: string;
  type?: string;
  upcoming?: boolean;
};