# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studio Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Current auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"
  /login:
    get:
      operationId: beginBrowserLogin
      tags: [auth]
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
...
  /logout:
    get:
      operationId: logoutBrowserSession
      tags: [auth]
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "302":
          description: Redirect to OIDC logout URL.
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"
  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        "200":
          description: Logout processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/{id}/log-outreach:
    post:
      operationId: logOutreach
      tags: [outreach]
      summary: Log an outreach for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutreachBody"
      responses:
        "201":
          description: Outreach logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outreach"
  /contacts/{id}/outreach:
    get:
      operationId: getContactOutreach
      tags: [outreach]
      summary: Get outreach history for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of outreach records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Outreach"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
...
    GetCurrentAuthUserResponse:
      type: object
      properties:
        isAuthenticated:
          type: boolean
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required:
        - isAuthenticated
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
      required:
        - id
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
      required:
        - code
        - code_verifier
        - redirect_uri
        - state
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        organization:
          type: string
        type:
          type: string
          description: "e.g. band_director, event_coordinator, venue, other"
        notes:
          type: string
        lastOutreachAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - createdAt
        - updatedAt
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        organization:
          type: string
        type:
          type: string
        notes:
          type: string
      required:
        - name
        - type
    Outreach:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        eventId:
          type: integer
        method:
          type: string
          description: "e.g. email, phone, in-person, text"
        notes:
          type: string
        outreachAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - method
        - outreachAt
        - createdAt
    CreateOutreachBody:
      type: object
      properties:
        method:
          type: string
        notes:
          type: string
        eventId:
          type: integer
        outreachAt:
          type: string
          format: date-time
      required:
        - method
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          description: "staff, intern"
        isActive:
          type: boolean
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        notes:
          type: string
      required:
        - name
        - role
...