# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  organization?: string;
  /** e.g. band_director, event_coordinator, venue, other */
  type: string;
  notes?: string;
  lastOutreachAt?: Date;
  createdAt: Date;
  updatedAt: Date;
}