# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  /** staff, intern */
  role: string;
  isActive: boolean;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}