# lib/api-zod/src/generated/types/outreach.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Outreach {
  id: number;
  contactId: number;
  eventId?: number;
  /** e.g. email, phone, in-person, text */
  method: string;
  notes?: string;
  outreachAt: Date;
  createdAt: Date;
}